// src/types/types.ts
type IAnnouncement = {
    $id: string;
    title: string;
    language: string;
    level: string;
    age: string;
    format: string;
    time: string;
    teacher: string;
    teacherId?: string;
    price: number;
    rating: number;
    createdAt: string;
    tags?: string[];
    description?: string;
  }
  